Re-structured everything
